fix ijon issue
