+++ ydb/core/tx/schemeshard/ut_export/ut_export.cpp
@@ -5,7 +5,7 @@
 #include <ydb/core/testlib/actors/block_events.h>
 #include <ydb/core/tx/datashard/datashard.h>
 #include <ydb/core/tx/schemeshard/schemeshard_billing_helpers.h>
-#include <ydb/core/tx/schemeshard/ut_helpers/auditlog_helpers.h>
+#include <ydb/core/testlib/audit_helpers/audit_helper.h>
 #include <ydb/core/tx/schemeshard/ut_helpers/helpers.h>
 #include <ydb/core/util/aws.h>
 #include <ydb/core/wrappers/s3_wrapper.h>
@@ -23,6 +23,10 @@ using namespace NSchemeShardUT_Private;
 using namespace NKikimr::NWrappers::NTestHelpers;
 
 using TTablesWithAttrs = TVector<std::pair<TString, TMap<TString, TString>>>;
+
+using TMemoryLogBackend = NKikimr::Tests::TMemoryLogBackend;
+using NKikimr::Tests::CreateTestAuditLogBackends;
+using NKikimr::Tests::FindAuditLine;
 
 namespace {
