+++ ydb/core/tx/schemeshard/ut_restore/ut_restore.cpp
@@ -10,7 +10,7 @@
 #include <ydb/core/tx/datashard/datashard.h>
 #include <ydb/core/tx/schemeshard/schemeshard_billing_helpers.h>
 #include <ydb/core/tx/schemeshard/schemeshard_private.h>
-#include <ydb/core/tx/schemeshard/ut_helpers/auditlog_helpers.h>
+#include <ydb/core/testlib/audit_helpers/audit_helper.h>
 #include <ydb/core/tx/schemeshard/ut_helpers/helpers.h>
 #include <ydb/core/util/aws.h>
 #include <ydb/core/wrappers/ut_helpers/s3_mock.h>
@@ -41,6 +41,10 @@ using namespace NKikimr::NWrappers::NTestHelpers;
 using namespace NKikimr;
 using namespace NKikimrSchemeOp;
 using namespace NSchemeShardUT_Private;
+
+using TMemoryLogBackend = NKikimr::Tests::TMemoryLogBackend;
+using NKikimr::Tests::CreateTestAuditLogBackends;
+using NKikimr::Tests::FindAuditLine;
 
 namespace {
